merge: feat/user + feat/project + feat/index.html
