+++ src/main/java/com/woowahan/moduchan/controller/ApiCategoryController.java
@@ -1,7 +1,5 @@
 package com.woowahan.moduchan.controller;
 
-import java.util.List;
-
 import com.woowahan.moduchan.domain.category.Category;
 import com.woowahan.moduchan.domain.project.Project;
 import com.woowahan.moduchan.dto.category.CategoryDTO;
@@ -9,8 +7,12 @@ import com.woowahan.moduchan.service.CategoryService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import springfox.documentation.annotations.Cacheable;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
 
 @RestController
 @RequestMapping("/api/categories")
