+++ src/main/java/com/woowahan/moduchan/controller/CategoryController.java
@@ -1,17 +1,10 @@
 package com.woowahan.moduchan.controller;
 
-import com.woowahan.moduchan.dto.category.CategoryDTO;
 import com.woowahan.moduchan.service.CategoryService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
-
-import java.util.List;
 
 @Controller
 @RequestMapping("/categories")
@@ -19,8 +12,8 @@ public class CategoryController {
     @Autowired
     private CategoryService categoryService;
 
-    @GetMapping("/{cid}")
-    public String showCategoryHTML(@PathVariable Long cid) {
-        return "category";
+    @GetMapping("")
+    public String collectProjects() {
+        return "/category";
     }
 }
