+++ src/main/java/com/woowahan/moduchan/domain/project/Project.java
@@ -1,5 +1,6 @@
 package com.woowahan.moduchan.domain.project;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.woowahan.moduchan.domain.category.Category;
 import com.woowahan.moduchan.domain.product.Product;
 import com.woowahan.moduchan.domain.user.NormalUser;
@@ -20,19 +21,9 @@ import java.util.List;
 @Getter
 @NoArgsConstructor
 public class Project extends BaseTimeEntity {
-    public enum STATUS {
-        DRAFT,
-        EVALUATING,
-        PUBLISHED,
-        REJECTED,
-        COMPLETE,
-        INCOMPLETE
-    }
-
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-
     /* 프로젝트 정보 */
     private String title;
     @Lob
@@ -42,22 +33,18 @@ public class Project extends BaseTimeEntity {
     private Long goalFundRaising;
     /* 시간 */
     private Date endAt;
-
     /*상태*/
     @Column(columnDefinition = "integer default 1")
     private STATUS status;
     private boolean deleted = false;
-
     /* 사람 */
     @ManyToOne
     @JoinColumn
     private NormalUser owner;
-
     @ManyToOne
     @JoinColumn
     @JsonIgnore
     private Category category;
-
     @OneToMany(cascade = CascadeType.ALL, fetch = FetchType.LAZY, orphanRemoval = true, mappedBy = "project")
     private List<Product> products = new ArrayList<>();
 
@@ -68,6 +55,15 @@ public class Project extends BaseTimeEntity {
         this.thumbnailUrl = thumbnailUrl;
         this.goalFundRaising = goalFundRaising;
         this.endAt = endAt;
+    }
+
+    public static Project from(ProjectDTO projectDTO) {
+        return new ProjectBuilder().title(projectDTO.getTitle())
+                .description(projectDTO.getDescription())
+                .endAt(projectDTO.getEndAt())
+                .goalFundRaising(projectDTO.getGoalFundRaising())
+                .thumbnailUrl(projectDTO.getThumbnailUrl())
+                .build();
     }
 
     public Project addCategory(Category category) {
@@ -84,16 +80,6 @@ public class Project extends BaseTimeEntity {
         productList.forEach(product -> this.products.add(product.erasePid().addProject(this)));
     }
 
-    public static Project from(ProjectDTO projectDTO) {
-        return new ProjectBuilder().title(projectDTO.getTitle())
-                .description(projectDTO.getDescription())
-                .endAt(projectDTO.getEndAt())
-                .goalFundRaising(projectDTO.getGoalFundRaising())
-                .thumbnailUrl(projectDTO.getThumbnailUrl())
-                .build();
-    }
-
-
     public Project updateProject(ProjectDTO projectDTO, Category category) {
         this.description = projectDTO.getDescription();
         this.endAt = projectDTO.getEndAt();
@@ -107,5 +93,14 @@ public class Project extends BaseTimeEntity {
     public void delete() {
         deleted = true;
         products.forEach(product -> product.delete());
+    }
+
+    public enum STATUS {
+        DRAFT,
+        EVALUATING,
+        PUBLISHED,
+        REJECTED,
+        COMPLETE,
+        INCOMPLETE
     }
 }
