+++ src/main/java/com/woowahan/moduchan/domain/user/User.java
@@ -1,11 +1,13 @@
 package com.woowahan.moduchan.domain.user;
 
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
 
 @MappedSuperclass
 @Getter
+@NoArgsConstructor
 public abstract class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
