+++ src/main/java/com/woowahan/moduchan/service/CategoryService.java
@@ -7,7 +7,6 @@ import com.woowahan.moduchan.repository.CategoryRepository;
 import com.woowahan.moduchan.repository.ProjectRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
