+++ src/main/java/com/woowahan/moduchan/support/BaseTimeEntity.java
@@ -6,7 +6,6 @@ import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import javax.persistence.EntityListeners;
 import javax.persistence.MappedSuperclass;
-import java.time.LocalDateTime;
 import java.util.Date;
 
 @Getter
