--- src/main/resources/static/category.html
@@ -1,23 +1,0 @@
-<html>
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport"
-          content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>Document</title>
-    <link rel="stylesheet" href="/css/category.css"/>
-    <script src="/js/category.js"></script>
-    <link href="https://fonts.googleapis.com/css?family=Roboto" rel="stylesheet">
-</head>
-<body>
-<div class="content">
-    <ul class="categories">
-
-    </ul>
-
-    <ul class="projects">
-
-    </ul>
-</div>
-</body>
-</html>
