--- src/main/resources/static/index.html
@@ -1,19 +1,0 @@
-<html>
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport"
-          content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>Document</title>
-    <link rel="stylesheet" href="/css/category.css"/>
-    <link href="https://fonts.googleapis.com/css?family=Roboto" rel="stylesheet">
-    <script src="/js/category.js"></script>
-</head>
-<body>
-<div class="promotion_div">
-    <ul class="promotion_ul">
-
-    </ul>
-</div>
-</body>
-</html>
