--- src/main/resources/static/products.html
@@ -1,18 +1,0 @@
-<html>
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport"
-          content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>Document</title>
-</head>
-<body>
-
-<div>
-    <ul class="categories_ul">
-        <li class="category_li"></li>
-    </ul>
-</div>
-
-</body>
-</html>
