+++ src/main/resources/templates/base.html
@@ -26,7 +26,7 @@
             </a>
         </h1>
         <div class="navigation-project">
-            <a href="/">
+            <a href="/categories">
                 <span class="collect-project">프로젝트 모아보기</span>
             </a>
             <a href="/">
