+++ src/main/resources/templates/category.html
@@ -1,0 +1,19 @@
+{{> base}}
+
+{{#block "header"}}
+<link rel="stylesheet" href="/css/category.css"/>
+{{/block}}
+
+{{#block "contents"}}
+<div class="content">
+    <ul class="categories">
+    </ul>
+    <ul class="projects">
+    </ul>
+</div>
+{{/block}}
+
+{{#block "footer"}}
+<script src="/js/common.js"></script>
+<script src="/js/category.js"></script>
+{{/block}}
